fix: post-restructuring adjustments
