+++ connector-runtime/connector-runtime-spring/src/main/java/io/camunda/connector/runtime/inbound/lifecycle/ActiveInboundConnector.java
@@ -18,7 +18,7 @@ package io.camunda.connector.runtime.inbound.lifecycle;
 
 import io.camunda.connector.api.inbound.InboundConnectorExecutable;
 import io.camunda.connector.impl.inbound.InboundConnectorProperties;
-import io.camunda.connector.runtime.util.inbound.InboundConnectorContextImpl;
+import io.camunda.connector.runtime.core.inbound.InboundConnectorContextImpl;
 
 public record ActiveInboundConnector(
     InboundConnectorExecutable executable,
