+++ connector-runtime/connector-runtime-spring/src/main/java/io/camunda/connector/runtime/outbound/jobhandling/SpringConnectorJobHandler.java
@@ -19,9 +19,9 @@ package io.camunda.connector.runtime.outbound.jobhandling;
 import io.camunda.connector.api.error.BpmnError;
 import io.camunda.connector.api.outbound.OutboundConnectorFunction;
 import io.camunda.connector.impl.outbound.OutboundConnectorConfiguration;
-import io.camunda.connector.runtime.util.outbound.ConnectorJobHandler;
-import io.camunda.connector.runtime.util.outbound.ConnectorResult;
-import io.camunda.connector.runtime.util.secret.SecretProviderAggregator;
+import io.camunda.connector.runtime.core.outbound.ConnectorJobHandler;
+import io.camunda.connector.runtime.core.outbound.ConnectorResult;
+import io.camunda.connector.runtime.core.secret.SecretProviderAggregator;
 import io.camunda.zeebe.client.api.command.CompleteJobCommandStep1;
 import io.camunda.zeebe.client.api.command.FinalCommandStep;
 import io.camunda.zeebe.client.api.response.ActivatedJob;
