+++ connector-runtime/spring-boot-starter-camunda-connectors/src/test/java/io/camunda/connector/runtime/secret/SpringSecretProviderTest.java
@@ -17,9 +17,9 @@
 package io.camunda.connector.runtime.secret;
 
 import io.camunda.connector.runtime.app.TestConnectorRuntimeApplication;
+import io.camunda.connector.runtime.core.secret.SecretProviderAggregator;
 import io.camunda.connector.runtime.secret.providers.BarSpringSecretProvider;
 import io.camunda.connector.runtime.secret.providers.FooSpringSecretProvider;
-import io.camunda.connector.runtime.util.secret.SecretProviderAggregator;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
