+++ pom.xml
@@ -6,7 +6,7 @@
   <parent>
     <groupId>io.camunda.connector</groupId>
     <artifactId>connector-parent</artifactId>
-    <version>0.10.0-SNAPSHOT</version>
+    <version>0.10.0-rc1</version>
     <relativePath/>
   </parent>
 
@@ -19,7 +19,7 @@
   <inceptionYear>2022</inceptionYear>
 
   <properties>
-    <version.spring-zeebe>8.2.1</version.spring-zeebe>
+    <version.spring-zeebe>8.2.2</version.spring-zeebe>
     <version.operate-client>8.1.8</version.operate-client>
 
     <version.spring-boot>3.1.0</version.spring-boot>
